requirements.txt
streamlit
audio-recorder-streamlit
openai-whisper
torch
numpy
scipy
gtts
googletrans==4.0.0-rc1
